server/requirements.txt
aniso8601==10.0.0
blinker==1.8.2
click==8.1.8
Flask==3.0.3
Flask-Cors==5.0.0
Flask-RESTful==0.3.10
importlib_metadata==8.5.0
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==2.1.5
python-dotenv==1.0.1
pytz==2025.2
six==1.17.0
Werkzeug==3.0.6
zipp==3.20.2
